SwingConsolePane: extract ConsolePanel from SwingConsolePane
SwingConsolePane is responsible for lazy initialization and showing the
console window at correct time. ConsolePanel is responsible for holding a
text area to show console output.

ConsolePanel will become a starting point for implementing a LoggingPanel.
